Migrate from Integer to BigInteger for UserIDs

+++ SayaBot/modules/sql/feds_sql.py
@@ -2,7 +2,7 @@ import threading
 
 from SayaBot import dispatcher
 from SayaBot.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
+from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, BigInteger
 from telegram.error import BadRequest, Unauthorized
 
 
@@ -58,7 +58,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
     __tablename__ = "feds_settings"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     should_report = Column(Boolean, default=True)
 
     def __init__(self, user_id):
@@ -902,4 +902,5 @@ __load_all_feds_chats()
 __load_all_feds_banned()
 __load_all_feds_settings()
 __load_feds_subscriber()
+
